yandex_practicum/sprint_8/tasks/K/task_test.go
package main

import (
	"fmt"
	"io"
	"strings"
	"testing"
)

type TestItem struct {
	input  io.Reader
	output string
}

func TestTask(t *testing.T) {
	tasks := generateTasks()

	for i, v := range tasks {
		t.Run(fmt.Sprintf("Test %d", i+1), func(t *testing.T) {
			buf := strings.Builder{}
			Solution(v.input, &buf)
			if strings.TrimSpace(buf.String()) != v.output {
				t.Errorf("Неверный ответ решения!\nОтвет: \n%v \nВерно: \n%v", buf.String(), v.output)
			}
		})
	}
}

func generateTasks() []TestItem {
	return []TestItem{
		{strings.NewReader(`gggggbbb
bbef
`), "-1"},
		{strings.NewReader(`z
aaaaaaa
`), "1"},
		{strings.NewReader(`ccccz
aaaaaz
`), "0"},
	}
}
